% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivist.github-package.R
\docType{package}
\name{archivist-package}
\alias{archivist-package}
\title{Tools for Archiving, Managing and Sharing R Objects via GitHub}
\description{
GitHub integration extension of the \pkg{archivist} package - tool for dtoring, 
restoring and searching for R objects.
}
\details{
For more detailed information visit \pkg{archivist} \code{wiki} on 
\href{https://github.com/pbiecek/archivist/wiki}{Github}.
}
\author{
Marcin Kosinski [aut, cre] \email{m.p.kosinski@gmail.com} \cr
Przemyslaw Biecek [aut] \email{przemyslaw.biecek@gmail.com}
}
\seealso{
Other archivist: \code{\link{archivist-github-integration}},
  \code{\link{pushGitHubRepo}}
}

