% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHooksToPrintGitHub.R
\name{addHooksToPrintGitHub}
\alias{addHooksToPrintGitHub}
\title{Add \pkg{archivist} Hooks to \pkg{rmarkdown} HTML Report}
\usage{
addHooksToPrintGitHub(class = "ggplot", repoDir = aoptions("repoDir"),
  repo = aoptions("repo"), user = aoptions("user"), branch = "master",
  subdir = aoptions("subdir"))
}
\arguments{
\item{class}{A character containing a name of class (one or more) that should be archived.}

\item{repoDir}{A character containing a name of Local Repository.}

\item{repo}{A character containing a name of a Git repository on which the Repository is archived.
If \code{repo = NULL} then hooks will be added to files in local directories.}

\item{user}{A character containing a name of a Git user on whose account the \code{repo} is created.}

\item{branch}{A character containing a name of Git Repository's branch on which the Repository is archived. 
Default \code{branch} is \code{master}.}

\item{subdir}{A character containing a name of a directory on Git repository 
on which the Repository is stored. If the Repository is stored in main folder on Git repository, this should be set 
to \code{subdir = "/"} as default.}
}
\description{
\code{addHooksToPrintGitHub} adds an overloaded version of the print function for objects of selected class. 
The overloaded function will add all objects of selected class to the \link{Repository} and then will add hooks to the HTML report (generated in \pkg{rmarkdown}) for these objects (\code{artifacts} - \link{archivist-package}).
This is GitHub version of \code{addHooksToPrint} and it automatically stores artifacts on GitHub - see examples.
}
\note{
One can specify \code{userTags} as in \link{archive} for artifacts by adding \code{"tags"} attribute.
See note secion about that in \link{archive}.

Bug reports and feature requests can be sent to \href{https://github.com/pbiecek/archivist/issues}{https://github.com/pbiecek/archivist/issues}
}
\examples{

\dontrun{
# only in Rmd report, links to github repository
addHooksToPrint(class="ggplot", repoDir = "arepo",
repo="graphGallery", user="pbiecek")
# only in Rmd report, links to local files
addHooksToPrint(class="ggplot", repoDir = "arepo",
repo=NULL)
}
}
\author{
Przemyslaw Biecek (\link{addHooksToPrint}), \email{przemyslaw.biecek@gmail.com} \cr
Marcin Kosinski (\link{addHooksToPrintGitHub}), \email{m.p.kosinski@gmail.com}
}
\seealso{
Other archivist: \code{\link{archivist-github-integration}},
  \code{\link{archivist-package}},
  \code{\link{pushGitHubRepo}}
}

