% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivist.github-package.R
\docType{package}
\name{archivist-package}
\alias{archivist-package}
\title{Tools for Archiving, Managing and Sharing R Objects via GitHub}
\description{
GitHub integration extension of the \pkg{archivist} package - tool for storing, 
restoring and searching for R objects. More about \pkg{archivis.github} can be found on 
\href{http://marcinkosinski.github.io/archivist.github/}{marcinkosinski.github.io/archivist.github/}
}
\details{
For more detailed information visit \pkg{archivist} \code{wiki} on 
\href{https://github.com/pbiecek/archivist/wiki}{Github}.
}
\note{
Bug reports and feature requests can be sent to \href{https://github.com/MarcinKosinski/archivist.github/issues}{https://github.com/MarcinKosinski/archivist.github/issues}
}
\author{
Marcin Kosinski [aut, cre] \email{m.p.kosinski@gmail.com} \cr
Przemyslaw Biecek [aut] \email{przemyslaw.biecek@gmail.com}
}
\seealso{
Other archivist: \code{\link{addHooksToPrintGitHub}},
  \code{\link{archivist-github-integration}},
  \code{\link{pushGitHubRepo}}
}

