% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivist.github-package.R
\docType{package}
\name{archivist.github-package}
\alias{archivist.github-package}
\title{Tools for Archiving, Managing and Sharing R Objects via GitHub}
\description{
GitHub integration extension of the \pkg{archivist} package - tool for storing, 
restoring and searching for R objects. More about \pkg{archivis.github} can be found on 
\href{http://marcinkosinski.github.io/archivist.github/}{marcinkosinski.github.io/archivist.github/}
}
\details{
More about \pkg{archivis.github} can be found on 
\href{http://marcinkosinski.github.io/archivist.github/}{marcinkosinski.github.io/archivist.github/} 
and about \pkg{archivist} in posts' history on \href{http://pbiecek.github.io/archivist/Posts.html}{http://pbiecek.github.io/archivist/Posts.html}#'
}
\note{
Bug reports and feature requests can be sent to \href{https://github.com/MarcinKosinski/archivist.github/issues}{https://github.com/MarcinKosinski/archivist.github/issues}
}
\author{
Marcin Kosinski [aut, cre] \email{m.p.kosinski@gmail.com} \cr
Przemyslaw Biecek [aut] \email{przemyslaw.biecek@gmail.com}
}
\seealso{
Other archivist.github: \code{\link{archive}},
  \code{\link{authoriseGitHub}},
  \code{\link{cloneGitHubRepo}},
  \code{\link{createGitHubRepo}},
  \code{\link{deleteGitHubRepo}},
  \code{\link{pushGitHubRepo}}
}

