% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authoriseGitHub.R
\name{authoriseGitHub}
\alias{authoriseGitHub}
\title{Authorise with GitHub API}
\usage{
authoriseGitHub(ClientID, ClientSecret)
}
\arguments{
\item{ClientID}{A 20 characters long string with Client ID. See https://github.com/settings/applications/ for more details.}

\item{ClientSecret}{A 40 characters long string with Client Secret. See https://github.com/settings/applications/ for more details.}
}
\description{
\code{authoriseGitHub} is function that performes OAuth authorisation with GitHub API 
and stores resulting token in the \code{github_token} variable.
In order to authorise your app you need ClinetID and ClientSecret.
They can be found here: https://github.com/settings/applications/new
}
\examples{
\dontrun{
## GitHub version
authoriseGitHub(ClientID, ClientSecret)
}
}
\author{
Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}
}
\seealso{
Other archivist.github: \code{\link{archive}},
  \code{\link{cloneGitHubRepo}},
  \code{\link{createGitHubRepo}},
  \code{\link{deleteGitHubRepo}}
}

